Converting docker to podman and fixing the pod forward container
